+++ apirest.php
@@ -40,7 +40,6 @@
 use Glpi\Cache\CacheManager;
 
 define('GLPI_ROOT', __DIR__);
-define('DO_NOT_CHECK_HTTP_REFERER', 1);
 ini_set('session.use_cookies', 0);
 
 include_once(GLPI_ROOT . "/inc/based_config.php");