+++ tools/getsearchoptions.php
@@ -33,8 +33,6 @@
  * ---------------------------------------------------------------------
  */
 
-define('DO_NOT_CHECK_HTTP_REFERER', 1);
-
 // Ensure current directory when run from crontab
 chdir(__DIR__);